feat(send): unify asset sends with offchain transaction builder
Move generic asset-send packet construction into ark_core::send, embed assets directly in VtxoInput, and remove the separate send_asset flow. Update clients and e2e tests to use the unified send API.
