+++ e2e-tests/tests/e2e_continue_pending_tx.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::unwrap_used)]
 
 use ark_core::coin_select::select_vtxos;
+use ark_core::send::VtxoInput;
 use bitcoin::key::Secp256k1;
 use bitcoin::Amount;
 use common::init_tracing;
@@ -58,12 +59,12 @@ pub async fn e2e_continue_pending_tx() {
 
     let selected = select_vtxos(spendable, send_amount, alice.server_info.dust, true).unwrap();
 
-    let vtxo_inputs: Vec<ark_core::send::VtxoInput> = selected
+    let vtxo_inputs: Vec<VtxoInput> = selected
         .into_iter()
         .map(|coin| {
             let vtxo = script_pubkey_to_vtxo_map.get(&coin.script_pubkey).unwrap();
             let (forfeit_script, control_block) = vtxo.forfeit_spend_info().unwrap();
-            ark_core::send::VtxoInput::new(
+            VtxoInput::new(
                 forfeit_script,
                 None,
                 control_block,
@@ -71,6 +72,7 @@ pub async fn e2e_continue_pending_tx() {
                 vtxo.script_pubkey(),
                 coin.amount,
                 coin.outpoint,
+                coin.assets,
             )
         })
         .collect();
